requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
Django==3.0.1
djangorestframework==3.11.0
idna==2.8
inscriptis==1.0
lxml==4.4.2
Pillow==6.2.1
pytz==2019.3
requests==2.22.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.7
gunicorn==19.9.0
whitenoise==4.1.2
django-heroku
